Added Flocking
Added some flocking.  Vehicle.cs is the script for the flocker, and it
extends AStarAgent.cs, so it ultilizes the pathfinding methods so it has
somewhere to flock to.

Also added is a GameManager class which contains all the weighting
parameters for the flocking, as well as the distances, speeds and
forces, etc.

**Obstacle avoidance is only a method stub for now, and will not work.
In addition, the code is fairly messy and may not be well-commented.  I
would like the single a* agent to also be in a derived script, to clean
things up a bit.